+++ packages/interfaces/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated the minimum version of `@tvkitchen/base-constants` to `1.1.1`.
 - Removed the unnecessarily required `emitResult` method from `IAppliance`.
+- Made the `IAppliance` `getInputTypes` and `getOutputTypes` methods static.
 
 ### Added
 - Create `isIAppliance` duck type method to IAppliance.